+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: aben-ham <aben-ham@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/16 14:50:08 by aben-ham          #+#    #+#             */
-/*   Updated: 2021/11/20 19:08:33 by aben-ham         ###   ########.fr       */
+/*   Updated: 2021/11/22 17:32:20 by aben-ham         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (*src != 0 && dstsize > 1)
