+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: aben-ham <aben-ham@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/18 10:16:22 by aben-ham          #+#    #+#             */
-/*   Updated: 2021/11/20 19:05:18 by aben-ham         ###   ########.fr       */
+/*   Updated: 2021/11/22 17:33:18 by aben-ham         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ static size_t	min(size_t a, size_t b)
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	char	*res;
-	int		substr_len;
+	size_t	substr_len;
 
 	if (!s)
 		return (NULL);
